Extract `calcCompletionDate` and refactor
