+++ src/core/production-line.ts
@@ -1,0 +1,37 @@
+export function calcCompletionDate(line: PrunApi.ProductionLine, order: PrunApi.ProductionOrder) {
+  if (!order.duration) {
+    return undefined;
+  }
+
+  if (order.completion) {
+    return order.completion.timestamp;
+  }
+
+  const capacity = line.capacity;
+  if (capacity === 0) {
+    return undefined;
+  }
+  const queue: number[] = [];
+
+  for (const lineOrder of line.orders) {
+    if (!lineOrder.duration) {
+      return undefined;
+    }
+    if (lineOrder.completion) {
+      // Order has started
+      queue.push(lineOrder.completion.timestamp);
+    } else if (queue.length < capacity) {
+      // Order has not started but there's capacity to start it
+      queue.push(Date.now() + lineOrder.duration.millis);
+    } else {
+      // Order has not started
+      queue.sort();
+      queue.push(queue.shift()! + lineOrder.duration.millis);
+    }
+    if (lineOrder === order) {
+      return queue.pop();
+    }
+  }
+
+  return undefined;
+}
