--- src/features/basic/prodq-order-eta copy.module.css
@@ -1,3 +1,0 @@
-.thCompletion {
-  width:min-content;
-}
