+++ src/features/basic/prodq-order-eta.module.css
@@ -1,0 +1,3 @@
+.thCompletion {
+  width: min-content;
+}
