+++ src/features/basic/prodq-order-eta.ts
@@ -5,25 +5,26 @@ import { timestampEachMinute } from '@src/utils/dayjs';
 import { createReactiveDiv } from '@src/utils/reactive-element';
 import { keepLast } from '@src/utils/keep-last';
 import { applyScopedCssRule } from '@src/infrastructure/prun-ui/refined-prun-css';
-import classes from './prodq-order-eta copy.module.css';
+import classes from './prodq-order-eta.module.css';
+import { calcCompletionDate } from '@src/core/production-line';
 
 function onTileReady(tile: PrunTile) {
-  if (!tile.parameter) {
-    return;
-  }
-
   subscribe($$(tile.anchor, C.ProductionQueue.table), table => {
     subscribe($$(table, 'tr'), order => {
-      if (_$(order, 'th')) {
+      const orderId = refPrunId(order);
+      if (!orderId.value) {
         return;
       }
-      onOrderSlotReady(order.children[5] as HTMLElement, order, tile.parameter!);
+      onOrderSlotReady(order.children[5] as HTMLElement, orderId, tile.parameter!);
     });
   });
 }
 
-function onOrderSlotReady(slot: HTMLElement, prunIdSlot: HTMLElement, siteId: string) {
-  const orderId = refPrunId(prunIdSlot);
+function onOrderSlotReady(
+  slot: HTMLElement,
+  orderId: globalThis.Ref<string | null, string | null>,
+  siteId: string,
+) {
   const completion = computed(() => {
     const line = productionStore.getById(siteId);
     for (const order of line?.orders ?? []) {
@@ -42,44 +43,6 @@ function onOrderSlotReady(slot: HTMLElement, prunIdSlot: HTMLElement, siteId: st
   });
   const div = createReactiveDiv(slot, eta);
   keepLast(slot, () => slot, div);
-}
-
-function calcCompletionDate(line: PrunApi.ProductionLine, order: PrunApi.ProductionOrder) {
-  if (!order.duration) {
-    return undefined;
-  }
-
-  if (order.completion) {
-    return order.completion.timestamp;
-  }
-
-  const capacity = line.capacity;
-  if (capacity === 0) {
-    return undefined;
-  }
-  const queue: number[] = [];
-
-  for (const lineOrder of line.orders) {
-    if (!lineOrder.duration) {
-      return undefined;
-    }
-    if (lineOrder.completion) {
-      // Order has started
-      queue.push(lineOrder.completion.timestamp);
-    } else if (queue.length < capacity) {
-      // Order has not started but there's capacity to start it
-      queue.push(Date.now() + lineOrder.duration.millis);
-    } else {
-      // Order has not started
-      queue.sort();
-      queue.push(queue.shift()! + lineOrder.duration.millis);
-    }
-    if (lineOrder === order) {
-      return queue.pop();
-    }
-  }
-
-  return undefined;
 }
 
 function init() {
